+++ lib/dart_quill_table_better.dart
@@ -1,0 +1,7 @@
+library;
+
+export 'src/table_better/config/config.dart';
+export 'src/table_better/formats/table.dart';
+export 'src/table_better/language/language.dart';
+export 'src/table_better/modules/clipboard.dart';
+export 'src/table_better/register.dart';